[#98] style: rename projections
