+++ src/main/java/com/nobrain/bookmarking/domain/bookmark_tag/dto/projection/BookmarkTagProjection.java
@@ -1,4 +1,4 @@
-package com.nobrain.bookmarking.domain.bookmark_tag.dto;
+package com.nobrain.bookmarking.domain.bookmark_tag.dto.projection;
 
 import com.nobrain.bookmarking.domain.bookmark.entity.Bookmark;
 import com.nobrain.bookmarking.domain.tag.entity.Tag;
@@ -6,12 +6,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-public class BookmarkTagDto {
+public class BookmarkTagProjection {
 
     @Getter
     @Builder
     @AllArgsConstructor
-    public static class BookmarkAndTagId {
+    public static class BookmarkAndTag {
         private Tag tag;
         private Bookmark bookmark;
     }
