+++ src/main/java/com/nobrain/bookmarking/domain/user/service/UserService.java
@@ -7,7 +7,7 @@ import com.nobrain.bookmarking.domain.user.entity.User;
 import com.nobrain.bookmarking.domain.user.exception.UserNotCorrectPasswordException;
 import com.nobrain.bookmarking.domain.user.exception.UserNotFoundException;
 import com.nobrain.bookmarking.domain.user.repository.UserRepository;
-import com.nobrain.bookmarking.domain.user.repository.mapping.UserInfo;
+import com.nobrain.bookmarking.domain.user.repository.projection.UserInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
